requirements.txt
Flask
Flask-CORS
SpeechRecognition
gTTS
google-generativeai